moved client related tests to proper place. fixed reading quer log
